requirements.txt
beautifulsoup4==4.13.4
cloudscraper==1.2.71
fastapi==0.115.12
groq==0.24.0
langchain==0.3.25
langchain-groq
pydantic==2.11.4
pymilvus==2.5.6
pymilvus_model==0.3.1
pytest==8.3.5
python-dotenv==1.1.0
sentence_transformers==3.4.1
tqdm==4.67.1
uvicorn==0.34.2